% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastdfe.R
\name{fastdfe_is_installed}
\alias{fastdfe_is_installed}
\title{Check if the \code{fastdfe} Python module is installed}
\usage{
fastdfe_is_installed()
}
\value{
Logical \code{TRUE} if the \code{fastdfe} Python module is installed, otherwise \code{FALSE}.
}
\description{
This function uses the reticulate package to verify if the \code{fastdfe} Python
module is currently installed.
}
\examples{
\dontrun{
is_installed()  # Returns TRUE or FALSE based on the installation status of fastdfe
}

}
