% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastdfe.R
\name{load_fastdfe}
\alias{load_fastdfe}
\title{Load the fastdfe library and associated visualization functions}
\usage{
load_fastdfe(install = FALSE)
}
\arguments{
\item{install}{A logical. If TRUE, the function will attempt to run install_fastdfe().}
}
\value{
A reference to the 'fastdfe' Python library loaded through reticulate.
This reference can be used to access 'fastdfe' functionalities.
}
\description{
This function imports the Python package 'fastdfe' using the reticulate package
and then configures it to work seamlessly with R, overriding some of the default
visualization functions with custom R-based ones. This function also ensures
that required R libraries are loaded for visualization.
}
\examples{
\dontrun{
load_fastdfe(install = TRUE)
# now you can use fastdfe functionalities as per its API
}

}
\seealso{
\link[reticulate]{import} for importing Python modules in R.
}
