% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastdfe.R
\name{install_fastdfe}
\alias{install_fastdfe}
\title{Install the \code{fastdfe} Python module}
\usage{
install_fastdfe(version = NULL, force = FALSE, silent = FALSE)
}
\arguments{
\item{version}{A character string specifying the version of the \code{fastdfe} module
to install. Default is \code{NULL} which will install the latest version.}

\item{force}{Logical, if \code{TRUE} it will force the reinstallation of the \code{fastdfe} module
even if it's already available. Default is \code{FALSE}.}

\item{silent}{Logical, if \code{TRUE} it will suppress the message about \code{fastdfe} being
already installed. Default is \code{FALSE}.}
}
\value{
Invisible \code{NULL}.
}
\description{
This function checks if the \code{fastdfe} Python module is available.
If not, or if the \code{force} argument is TRUE, it installs it via pip.
If the \code{silent} argument is set to TRUE, the function will not output a
message when the module is already installed.
}
\examples{
\dontrun{
install_fastdfe()  # Installs the latest version of fastdfe
install_fastdfe("1.0.0")  # Installs version 1.0.0 of fastdfe
install_fastdfe(force = TRUE)  # Reinstalls the fastdfe module
}

}
